app_streamlit/requirements.txt
streamlit
pandas
scikit-learn
joblib
pyyaml
